Update newsletters and fix nav links
