+++ menuItems.ts
@@ -67,8 +67,8 @@ export const menuItems: MenuItems = [
         title: 'Updates',
         url: '/updates',
         submenu: [
-          { title: 'Blog', url: '/blogs' },
-          { title: 'Newsletters', url: '/newsletters' }
+          { title: 'Blog', url: 'blogs' },
+          { title: 'Newsletters', url: 'newsletters' }
         ]
     },   
     {
