+++ pages/updates/newsletters.tsx
@@ -13,6 +13,12 @@ export default function Newsletters() {
       <div className="newsletters-list">
         <ul>
           <li>
+            <a href="/newsletterApril2025.pdf" target="_blank" className="text-lg text-blue-500 underline">2025 Newsletter - Vol 2</a>
+          </li>
+          <li>
+            <a href="/newsletterJan2025.pdf" target="_blank" className="text-lg text-blue-500 underline">2025 Newsletter - Vol 1</a>
+          </li>
+          <li>
             <a href="/newsletter.pdf" target="_blank" className="text-lg text-blue-500 underline">2024 Newsletter - Vol 1</a>
           </li>
         </ul>
